Release wakelock on power-off
This makes sure the bluesleep wakelock is released upon power-off.
Previously this was explicitely done in the HAL, but the calling code
should not need to be aware of the internal state and callbacks
shouldn't be triggered on power-off to avoid race conditions.

Bug: 36358814
Test: manual
Change-Id: I157b4c718ad9520dd7ce0b84990e602728b1e636
(cherry picked from commit b31ab240a11a8ce5809d6abeba3c05829a155da2)